+++ src/main/java/com/revolvingmadness/sculk/language/parser/Parser.java
@@ -552,9 +552,15 @@ public class Parser {
 
         this.consume(TokenType.RIGHT_PARENTHESIS);
 
-        this.consume(TokenType.RIGHT_ARROW);
+        String returnType;
 
-        String returnType = (String) this.consume(TokenType.IDENTIFIER).value;
+        if (this.current(TokenType.RIGHT_ARROW)) {
+            this.consume(TokenType.RIGHT_ARROW);
+
+            returnType = (String) this.consume(TokenType.IDENTIFIER).value;
+        } else {
+            returnType = "Null";
+        }
 
         List<StatementNode> body = this.parseBody();